+++ math/util.h
@@ -5,21 +5,21 @@
 #include <random>
 
 // Constants
-const float infinity = std::numeric_limits<float>::infinity();
-const float pi = 3.1415926535897932385;
+const double infinity = std::numeric_limits<double>::infinity();
+const double pi = 3.1415926535897932385;
 
 // Utility functions
-inline float deg2rad(float deg) {
+inline double deg2rad(double deg) {
 	return deg * pi / 180.0;
 }
 
-static std::uniform_real_distribution<float> distribution(0.0, 1.0);
+static std::uniform_real_distribution<double> distribution(0.0, 1.0);
 static std::mt19937 generator;
 
-inline float randf() {
+inline double randf() {
 	return distribution(generator);
 }
-inline float randf(float min, float max) {
+inline double randf(double min, double max) {
 	return min + (max - min) * randf();
 }
