+++ util/color.h
@@ -7,7 +7,7 @@
 
 using color = vec3;
 
-inline float apply_gamma(float f) {
+inline double apply_gamma(double f) {
     if (f > 0) {
         return std::sqrt(f);
     }
@@ -17,9 +17,9 @@ inline float apply_gamma(float f) {
 void write_color(std::ofstream& file, const color& col) {
     // Red from left to right
     // Green from top to bottom
-    float r = apply_gamma(col.x);
-    float g = apply_gamma(col.y);
-    float b = apply_gamma(col.z);
+    double r = apply_gamma(col.x);
+    double g = apply_gamma(col.y);
+    double b = apply_gamma(col.z);
 
     const interval intensity{ 0.0, 1.0 };
     int ir = (int)(intensity.clamp(r) * 255.999);
